+++ include/vigra/multi_convolution.hxx
@@ -561,7 +561,7 @@ gaussianGradientMultiArray(SrcIterator si, SrcShape const & shape, SrcAccessor s
         if(shape[k] <=0)
             return;
 
-    vigra_precondition(N == dest.size(di),
+    vigra_precondition(N == (int)dest.size(di),
         "gaussianGradientMultiArray(): Wrong number of channels in output array.");
 
     vigra_precondition(sigma > 0.0, "gaussianGradientMultiArray(): Scale must be positive.");
@@ -672,7 +672,7 @@ symmetricGradientMultiArray(SrcIterator si, SrcShape const & shape, SrcAccessor 
         if(shape[k] <=0)
             return;
 
-    vigra_precondition(N == dest.size(di),
+    vigra_precondition(N == (int)dest.size(di),
         "symmetricGradientMultiArray(): Wrong number of channels in output array.");
 
     Kernel1D<KernelType> filter;
@@ -902,7 +902,7 @@ hessianOfGaussianMultiArray(SrcIterator si, SrcShape const & shape, SrcAccessor 
         if(shape[k] <=0)
             return;
 
-    vigra_precondition(M == dest.size(di),
+    vigra_precondition(M == (int)dest.size(di),
         "hessianOfGaussianMultiArray(): Wrong number of channels in output array.");
 
     vigra_precondition(sigma > 0.0, "hessianOfGaussianMultiArray(): Scale must be positive.");
@@ -1058,7 +1058,7 @@ structureTensorMultiArray(SrcIterator si, SrcShape const & shape, SrcAccessor sr
         if(shape[k] <=0)
             return;
 
-    vigra_precondition(M == dest.size(di),
+    vigra_precondition(M == (int)dest.size(di),
         "structureTensorMultiArray(): Wrong number of channels in output array.");
 
     vigra_precondition(innerScale > 0.0 && outerScale >= 0.0,
